22.12.4 2:04
add ShopPanelBaseRequest,
ShopPanelInfo 93% -> 98% ?

+++ Assets/Scripts/Info/Entity/PlayerInfo.cs
@@ -122,8 +122,8 @@ public class PlayerInfo : EntityInfo
     {
         get { return diamond; }
     }
-    [SerializeField] long coin;
-    public long Coin
+    [SerializeField] float coin;
+    public float Coin
     {
         get { return coin; }
     }
